refactor: :recycle: move type files to lib folder
they were not getting included in the dist folder
